requirements.txt
flask==2.3.3
cryptography==41.0.7
redis==5.0.1
pymongo==4.6.0
elasticsearch==8.11.0
gitpython==3.1.40
opencv-python==4.8.1.78
folium==0.14.0
geopandas==0.13.2
asyncio
requests==2.31.0
pytest==7.4.3
